complete Main page filter activities function
